frontend/requirements.txt
gradio==3.0
psycopg2-binary==2.9.3
pandas==1.4.2
